channel_simulator/requirements.txt
Pillow>=10.0.0
pytest>=7.0.0
numpy>=1.20.0
jpeglib>=0.14.0
reedsolo>=1.7.0
